refactor(fetcher): split HuggingFace model fetching modules
Move fetcher responsibilities into focused modules for HF API orchestration, model parsing, parameter normalization, GGUF extraction, sliding-window metadata, and cache serialization.

Keep whichllm.models.fetcher as a compatibility shim so existing public and test imports continue to work.

Refs #41
